pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "cliqq"
version = "0.0.1"
# check using tox + vermin or alternatives
dependencies = [
  "prompt_toolkit>=3.0.0",
  "psutil>=5.9.0",
  "python-dotenv>=1.0.0",
  "openai>=1.0.0",
  # pytest
  # psutil
]
authors = [{ name = "Tobi Adesanya", email = "tobijadesanya@gmail.com" }]
description = "a simple command line chat assistant"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/tobiia/cliqq-cli"

[project.entry-points."spam.magical"]
tomatoes = "spam:main_tomatoes"
